+++ ca/ca-dbtool/src/main/java/org/xipki/pki/ca/dbtool/diffdb/io/CaEntryContainer.java
@@ -61,7 +61,7 @@ public class CaEntryContainer {
         }
     }
 
-    public void addDigestEntry(final int caId, final int id, final DbDigestEntry reportEntry)
+    public void addDigestEntry(final int caId, final long id, final DbDigestEntry reportEntry)
     throws IOException, InvalidDataObjectException {
         CaEntry ce = caEntryMap.get(caId);
         if (ce == null) {
