+++ ca/ca-server/src/main/java/org/xipki/pki/ca/server/impl/scep/Scep.java
@@ -521,7 +521,7 @@ public class Scep {
                         encodedRequest = null;
                     }
                     if (encodedRequest != null) {
-                        int reqId = ca.addRequest(encodedRequest);
+                        long reqId = ca.addRequest(encodedRequest);
                         ca.addRequestCert(reqId, cert.getCert().getCertId());
                     }
                 }
